op: move impl function to op_impl.c
