+++ src/mpi/coll/op/op_create.c
@@ -24,78 +24,6 @@ int MPI_Op_create(MPI_User_function * user_fn, int commute, MPI_Op * op)
 #undef MPI_Op_create
 #define MPI_Op_create PMPI_Op_create
 
-#ifndef MPIR_OP_PREALLOC
-#define MPIR_OP_PREALLOC 16
-#endif
-
-/* Preallocated op objects */
-MPIR_Op MPIR_Op_builtin[MPIR_OP_N_BUILTIN];
-MPIR_Op MPIR_Op_direct[MPIR_OP_PREALLOC];
-
-MPIR_Object_alloc_t MPIR_Op_mem = { 0, 0, 0, 0, 0, 0, MPIR_OP,
-    sizeof(MPIR_Op),
-    MPIR_Op_direct,
-    MPIR_OP_PREALLOC,
-    NULL
-};
-
-#ifdef HAVE_CXX_BINDING
-void MPII_Op_set_cxx(MPI_Op op, void (*opcall) (void))
-{
-    MPIR_Op *op_ptr;
-
-    MPIR_Op_get_ptr(op, op_ptr);
-    op_ptr->language = MPIR_LANG__CXX;
-    MPIR_Process.cxx_call_op_fn = (void (*)(const void *, void *, int,
-                                            MPI_Datatype, MPI_User_function *)) opcall;
-}
-#endif
-#if defined(HAVE_FORTRAN_BINDING) && !defined(HAVE_FINT_IS_INT)
-/* Normally, the C and Fortran versions are the same, by design in the
-   MPI Standard.  However, if MPI_Fint and int are not the same size (e.g.,
-   MPI_Fint was made 8 bytes but int is 4 bytes), then the C and Fortran
-   versions must be distinquished. */
-void MPII_Op_set_fc(MPI_Op op)
-{
-    MPIR_Op *op_ptr;
-
-    MPIR_Op_get_ptr(op, op_ptr);
-    op_ptr->language = MPIR_LANG__FORTRAN;
-}
-#endif
-
-
-int MPIR_Op_create_impl(MPI_User_function * user_fn, int commute, MPI_Op * op)
-{
-    MPIR_Op *op_ptr;
-    int mpi_errno = MPI_SUCCESS;
-
-    op_ptr = (MPIR_Op *) MPIR_Handle_obj_alloc(&MPIR_Op_mem);
-    /* --BEGIN ERROR HANDLING-- */
-    if (!op_ptr) {
-        mpi_errno =
-            MPIR_Err_create_code(MPI_SUCCESS, MPIR_ERR_RECOVERABLE, __func__, __LINE__,
-                                 MPI_ERR_OTHER, "**nomem", "**nomem %s", "MPI_Op");
-        goto fn_fail;
-    }
-    /* --END ERROR HANDLING-- */
-
-    op_ptr->language = MPIR_LANG__C;
-    op_ptr->kind = commute ? MPIR_OP_KIND__USER : MPIR_OP_KIND__USER_NONCOMMUTE;
-    op_ptr->function.c_function = (void (*)(const void *, void *,
-                                            const int *, const MPI_Datatype *)) user_fn;
-    MPIR_Object_set_ref(op_ptr, 1);
-
-    MPIR_OBJ_PUBLISH_HANDLE(*op, op_ptr->handle);
-
-    MPID_Op_commit_hook(op_ptr);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 #endif /* MPICH_MPI_FROM_PMPI */
 
 /*@
